enhancement: add OTLP origin detection (#967)
## Summary
<!-- Please provide a brief summary about what this PR does.
This should help the reviewers give feedback faster and with higher
quality. -->
This PR adds a `OtlpOriginTagsResolver` in order to support getting the
origin tags for OTLP Metrics.

The tags come from values in `ResourceAttributes`, and the logic is
copied from
[here](https://github.com/DataDog/datadog-agent/blob/main/comp/otelcol/otlp/components/processor/infraattributesprocessor/common.go#L158).

Note, only container id, pod id, process id are added here. There are
some specific keys (i.e `AWSECSTaskARNKey`) that are missing in this PR
since `RawOrigin` does not support that currently, with the dogstatsd
source not using them.

## Change Type
- [ ] Bug fix
- [ ] New feature
- [ ] Non-functional (chore, refactoring, docs)
- [ ] Performance


## How did you test this PR?
<!-- Please how you tested these changes here -->

Sent some OTLP metrics with resource attributes populated with
`container_id` and made sure it showed up on datadog.

## References

<!-- Please list any issues closed by this PR. -->

<!--
- Closes: <issue link>
-->

<!-- Any other issues or PRs relevant to this PR? Feel free to list them
here. -->