enhancement: add new OTLP Relay component (#996)
## Summary
<!-- Please provide a brief summary about what this PR does.
This should help the reviewers give feedback faster and with higher
quality. -->
This pr adds a new `Relay` component that forwards raw bytes as is to
downstream components. This is needed so that we can create the OTLP
Relay component which will serve as a proxy for OTLP metrics/logs/traces
to the core-agent.

This pr also refactors some shared code used by the OTLP source, since
we still want to keep that code for now even though we aren't doing the
translation.

OTLP proxy mode is gated behind the `adp.otlp.proxy_mode` config.

Grpc Payload type has also been created since the trace-agent uses a
grpc server to listen for traces.

A following pr will be made to the datadog agent to ensure that there
aren't port conflicts since, the goal is for the user to send these OTLP
signals to ADP which will use the default agent ports.

## Change Type
- [ ] Bug fix
- [x] New feature
- [ ] Non-functional (chore, refactoring, docs)
- [ ] Performance

## How did you test this PR?
<!-- Please how you tested these changes here -->

Manually by sending OTLP metrics,logs,traces and making sure it worked.

## References

<!-- Please list any issues closed by this PR. -->

<!--
- Closes: <issue link>
-->

<!-- Any other issues or PRs relevant to this PR? Feel free to list them
here. --> 54a56303bd47375fd435806e7ed8365fe6409ba5