chore(dev): add new Make target/CI job for checking for unused dependencies (#1042)
## Summary

This PR adds a new Make target (and CI job) for checking for unused dependencies.

Pretty straightforward. We use [`cargo-machete`](https://github.com/bnjbvr/cargo-machete) which I've used by hand in the past and had instrumented various crates (the proto ones, in particular) to work with.. so this is just extending that to make it a proper part of the development process now.

I slightly reworked how we pull down tools via Cargo so that we use [`cargo-binstall`](https://github.com/cargo-bins/cargo-binstall), which just tries to use pre-compiled binaries directly from the upstream tool's source repository when possible, and in turn installs things much, much faster.

Finally, I removed all the unused deps that `cargo-machete` found so the CI should be green on this PR for this job. :)

## Change Type

- [ ] Bug fix
- [ ] New feature
- [x] Non-functional (chore, refactoring, docs)
- [ ] Performance

## How did you test this PR?

Ran `make check-unused-deps` locally to ensure it pulls down the `cargo-machete` tool _and_ properly detects unused dependencies.

## References

AGTMETRICS-393


Co-authored-by: toby.lawrence <toby.lawrence@datadoghq.com>