chore(datadog): add Datadog Traces encoder (#1054)
## Summary
<!-- Please provide a brief summary about what this PR does.
This should help the reviewers give feedback faster and with higher
quality. -->

This PR adds the trace encoder which receives a `Trace` event and
creates a `TracerPayload` to send to the forwarder as a HttpPayload

There also are a number of function/variables that are defined in the
private folder `source/otlp` that the Trace encoder requires. The shared
functions have been moved to a common location
`lib/saluki-components/src/common/otlp/util.rs`
## Change Type
- [ ] Bug fix
- [x] New feature
- [ ] Non-functional (chore, refactoring, docs)
- [ ] Performance


## How did you test this PR?
<!-- Please how you tested these changes here -->



## References

<!-- Please list any issues closed by this PR. -->

<!--
- Closes: <issue link>
-->

<!-- Any other issues or PRs relevant to this PR? Feel free to list them
here. -->