chore(dev): add basic Vale configuration (#1065)
## Summary

This PR adds a basic Vale configuration to the repository, in an opt-in fashion.

Vale is a linter for writing: it helps ensure that all different forms of text in a project -- prose, code comments, etc -- have consistent style, spelling, and so on. Saluki has a fairly specific "tone" when it comes to documentation, and we strive to ensure that our documentation is high-quality... but this is hard at scale and requires a lot of effort when reviewing contributions. This is the perfect use case for adding a tool that can do this laborious checking on our behalf.

I've added a basic Vale configuration to the repository, which means:

- we have new a Make target for running Vale against all relevant assets (`make check-docs`, which checks everything under `bin/`, `lib/`, and `docs/`)
- a basic Vale configuration that pulls from Google's development documentation style guide, as well as "write-good", a more prose-oriented ruleset
- some custom Saluki-specific rules, such as a dictionary of technical terms to allow and specific substitution rules (`Datadog` can't be `DataDog` or `datadog`, etc)

There are... a _ton_ of errors/warnings/suggestions with this configuration, and so we're not making any changes to start _enforcing_ usage of Vale as of this PR. Vale currently doesn't differentiate between doc (`///`) and non-doc (`//`) comments in Rust code, so it's applying these rules to non-doc comments which leads to the _vast_ majority of these errors/warnings/suggestions, and we don't want that. I opened an [issue](https://github.com/errata-ai/vale/issues/1060) upstream to figure out if it's possible to only lint against doc comments. Once there's some sort of resolution there, one way or the other, I'll work on actually trying to get Vale integrated as part of CI.

## Change Type
- [ ] Bug fix
- [ ] New feature
- [x] Non-functional (chore, refactoring, docs)
- [ ] Performance

## How did you test this PR?

N/A

## References

AGTMETRICS-393