chore: initial ADP vs Agent OTLP traces correctness test (#1073)
## Summary
<!-- Please provide a brief summary about what this PR does.
This should help the reviewers give feedback faster and with higher
quality. -->
This pr updates the existing OTLP traces correctness test configuration
so that we are now comparing the core-agent against ADP instead of agent
to agent.

Also includes some quick fixes caught in the correctness output.

Note: The test is going to be failing.
## Change Type
- [ ] Bug fix
- [ ] New feature
- [ ] Non-functional (chore, refactoring, docs)
- [ ] Performance


## How did you test this PR?
<!-- Please how you tested these changes here -->
`make test-correctness-otlp-traces`
## References

<!-- Please list any issues closed by this PR. -->

<!--
- Closes: <issue link>
-->

<!-- Any other issues or PRs relevant to this PR? Feel free to list them
here. -->